% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dsei_municipalities}
\alias{dsei_municipalities}
\title{Indigenous Sanitary Districts' municipalities}
\format{
A dataset with 516 rows and 3 columns
\describe{
\item{dsei}{Name of the Indigenous Sanitary District ('DSEI')}
\item{municipality_id_6}{Municipality's 6 digit code, according to IBGE}
\item{municipality_name}{Municipality's name}
}
}
\source{
Secretaria Especial de Saúde Indígena, SESAI. Lei de Acesso à
Informação
}
\usage{
dsei_municipalities
}
\description{
The dataset, which comes from the Brazilian open information legislation,
"Lei de Acesso à Informação" showcases in which municipality each Sanitary
District is located in. Different Sanitary Districts can be located in the
same municipality, e.g. Altamira municipality has 3 Sanitary Districts.
}
\keyword{datasets}
