% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dsei_population}
\alias{dsei_population}
\title{Indigenous Sanitary Districts' population}
\format{
A dataset with 350 rows and 3 columns
\describe{
\item{dsei}{Name of the Indigenous Sanitary District ('DSEI')}
\item{year}{Year of the observation}
\item{population}{Indigenous population living within the DSEI area, in
traditional settlements}
}
}
\source{
Secretaria Especial de Saúde Indígena, SESAI. Lei de Acesso à
Informação
}
\usage{
dsei_population
}
\description{
The dataset, which comes from the Brazilian open information legislation,
"Lei de Acesso à Informação" showcases indigenous population values
in each of the 34 Indigenous Sanitary Districts, the so called "DSEI". The
data ranges from 2010 to 2019.
}
\keyword{datasets}
