% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indigenous_covid_stats.R
\name{get_indigenous_covid_stats}
\alias{get_indigenous_covid_stats}
\title{Get indigenous covid stats}
\usage{
get_indigenous_covid_stats()
}
\value{
The function returns a tibble. It with 35 rows, each accounting for
one Sanitary District and one for the summarized amounts among all of them.
The tibble has 8 variables which characterize the covid situation in each of
the Districts.
}
\description{
This function web scrapes the official Brazilian website for covid among
indigenous people. The function has no parameters, and it only returns a
tibble with the most updated covid-19 stats, such as confirmed and suspect
cases, clinically cured patients and deaths caused by covid-19.
}
