% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indigenous_deaths}
\alias{indigenous_deaths}
\title{Indigenous reported deaths, from 2010 to 2019}
\format{
A dataset with 29349 rows and 12 columns
\describe{
\item{death_id}{Specific death identification code}
\item{dsei}{Name of the Indigenous Sanitary District ('DSEI')}
\item{base_pole_id}{Base pole in which the death was registered id}
\item{base_pole_name}{Base pole in which the death was registered name}
\item{municipality_id_6}{Municipality's 6 digit code, according to IBGE}
\item{municipality_name}{Municipality's name}
\item{state_abrev}{Brazilian state abbreviation}
\item{sex}{Biological sex of the deceased}
\item{age}{Age of the deceased}
\item{death_date}{Date of the death, Year-Month-Day}
\item{death_icd10}{Death's ICD10}
\item{death_cause}{Official cause of death, in portuguese}
}
}
\source{
Secretaria Especial de Saúde Indígena, SESAI. Lei de Acesso à
Informação
}
\usage{
indigenous_deaths
}
\description{
The dataset, which comes from the Brazilian open information legislation,
"Lei de Acesso à Informação" display indigenous reported deaths in Brazil.
Each observation consists in a different death.Only indigenous living in
traditional settlements are considered.
}
\keyword{datasets}
